Renamed files and directories to match dart convention
Reminder: use snake_case for files and directories names in lib/
