+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:directus/directus.dart';
 import 'package:flutter/material.dart';
-import 'package:pandasenp_flutter/Pages/Welcome.dart';
+import 'package:pandasenp_flutter/pages/welcome.dart';
 import 'package:pandasenp_flutter/directus/directus.dart';
 import 'package:pandasenp_flutter/env.dart';
