+++ lib/pages/login.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:pandasenp_flutter/Pages/Home.dart';
+import 'package:pandasenp_flutter/pages/home.dart';
+import 'package:pandasenp_flutter/pages/register.dart';
 
-class RegisterPage extends StatelessWidget {
+class LoginPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: const Color(0xFFEDECF2),
       appBar: AppBar(
-        title: const Text("Register"),
+        title: const Text("Login"),
         backgroundColor: Colors.black.withOpacity(0.7),
       ),
       body: SafeArea(
@@ -30,7 +31,7 @@ class RegisterPage extends StatelessWidget {
                 height: 30,
               ),
               const Text(
-                "Register with your email and password",
+                "Log in with your email and password",
                 style: TextStyle(
                   fontWeight: FontWeight.w500,
                   fontSize: 16,
@@ -42,37 +43,6 @@ class RegisterPage extends StatelessWidget {
               ),
 
               // Form
-
-              // Username
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 20.0),
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    border: Border.all(color: Colors.black),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  child: const Padding(
-                    padding: EdgeInsets.only(left: 20.0),
-                    child: TextField(
-                      decoration: InputDecoration(
-                        border: InputBorder.none,
-                        hintText: 'Username',
-                        hintStyle: TextStyle(
-                          color: Colors.grey,
-                        ),
-                      ),
-                      style: TextStyle(
-                        color: Colors.black,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-
-              const SizedBox(
-                height: 20,
-              ),
 
               // Email
               Padding(
@@ -134,38 +104,6 @@ class RegisterPage extends StatelessWidget {
               ),
 
               const SizedBox(
-                height: 20,
-              ),
-
-              // Re password
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 20.0),
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    border: Border.all(color: Colors.black),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  child: const Padding(
-                    padding: EdgeInsets.only(left: 20.0),
-                    child: TextField(
-                      obscureText: true,
-                      decoration: InputDecoration(
-                        border: InputBorder.none,
-                        hintText: 'Retype password',
-                        hintStyle: TextStyle(
-                          color: Colors.grey,
-                        ),
-                      ),
-                      style: TextStyle(
-                        color: Colors.black,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-
-              const SizedBox(
                 height: 30,
               ),
 
@@ -173,19 +111,60 @@ class RegisterPage extends StatelessWidget {
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
+                  // Login Button
+                  Container(
+                    height: 50,
+                    width: 150,
+                    decoration: BoxDecoration(
+                      color: Colors.black.withOpacity(0.7),
+                      borderRadius: BorderRadius.circular(20),
+                    ),
+                    child: ElevatedButton(
+                      onPressed: () {
+                        Navigator.push(context, MaterialPageRoute(builder: (context) => HomePage()));
+                      },
+                      style: ElevatedButton.styleFrom(
+                        primary: Colors.black.withOpacity(0.7),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30),
+                        ),
+                      ),
+                      child: const Text(
+                        "Login",
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 20,
+                        ),
+                      ),
+                    ),
+                  ),
+
+                  const SizedBox(
+                    width: 20,
+                  ),
+
                   // Register
                   Container(
                     height: 50,
                     width: 150,
                     decoration: BoxDecoration(
-                      color: Colors.black,
+                      color: Colors.white,
                       borderRadius: BorderRadius.circular(20),
                     ),
-                    child: const Center(
-                      child: Text(
+                    child: ElevatedButton(
+                      onPressed: () {
+                        Navigator.push(context, MaterialPageRoute(builder: (context) => RegisterPage()));
+                      },
+                      style: ElevatedButton.styleFrom(
+                        primary: Colors.white,
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30),
+                        ),
+                      ),
+                      child: const Text(
                         "Register",
                         style: TextStyle(
-                          color: Colors.white,
+                          color: Colors.black,
                           fontSize: 20,
                         ),
                       ),
